+++ packages/calcite-components-angular/projects/component-library/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.0-next.19](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.11.0-next.18...@esri/calcite-components-angular@2.11.0-next.19) (2024-07-23)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
+## [2.11.0-next.18](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.11.0-next.17...@esri/calcite-components-angular@2.11.0-next.18) (2024-07-22)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
+## [2.11.0-next.17](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.11.0-next.16...@esri/calcite-components-angular@2.11.0-next.17) (2024-07-22)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
 ## [2.11.0-next.16](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.11.0-next.15...@esri/calcite-components-angular@2.11.0-next.16) (2024-07-22)
 
 **Note:** Version bump only for package @esri/calcite-components-angular